+++ src/app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/Components/UnitViewSeoAccordion/index.ts
@@ -1,0 +1,1 @@
+export { UnitViewSeoAccordion } from "./UnitViewSeoAccordion";
